Verified Model (Call to API) functionality
The Model is called by the Presenter and do with retrofit the call to the API. It is an asyncronous call, so
the response is in a callback, the model send to the presenter the information, and the presenter log it

+++ app/src/main/java/com/vjsantojaca/openweather/model/api/WeatherCityResponse.kt
@@ -17,5 +17,5 @@ class ListWeatherResponse (
 )
 
 class MainWeatherResponse (
-    val temp : Int
+    val temp : Float
 )
